backend/pipelines/bmd_scraper/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bmd_scraper"
version = "0.1.0"
description = "Data pipeline for extracting and processing data from the Danish Bekæmpelsesmiddel Database"
requires-python = ">=3.9"
license = {text = "Proprietary"}
readme = "README.md"
dependencies = [
    "requests>=2.25.1",
    "beautifulsoup4>=4.9.3",
    "pandas>=1.3.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.1.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

[tool.setuptools]
packages = ["bronze", "silver"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100 